Main group 8 back fluxo1

+++ back-end/src/database/migrations/20230227195055-create-product.js
@@ -15,7 +15,7 @@ module.exports = {
       },
       price: {
         allowNull: false,
-        type: Sequelize.DOUBLE,
+        type: Sequelize.DECIMAL(4,2),
       },
       urlImage: {
         allowNull: false,
